requirements.txt
torch
transformers
accelerate
gradio
Pillow
